added item functions removed computed neededEXPItems computed
